third_party/blink/web_tests/fast/forms/select/menulist-popup-item-style-expected.txt

Tests fix for regression crbug.com/395613. Checks that option in optgroup is styled properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.isSelectPopupVisible(popup) is true
PASS internals.selectPopupItemStyleIsRtl(popup, 2) is true
PASS internals.isSelectPopupVisible(popup) is false
PASS internals.isSelectPopupVisible(popup) is true
PASS internals.selectPopupItemStyleFontHeight(popup, 2) is > 50
PASS successfullyParsed is true

TEST COMPLETE

Same ببببب
Same ببببب
